% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ann}
\alias{ann}
\title{ANN Model}
\usage{
ann(oy, ou, nrn, afc)
}
\arguments{
\item{oy}{Vector with the lags in the autoregressive terms}

\item{ou}{Vector with the lags in the exogenous terms}

\item{nrn}{a vector (with dimension = number of hidden layers) with the number of neurons in each hidden layers}

\item{afc}{activation functions of the neurons (see ? layer_dense)}
}
\value{
Object representing the ANN-NARX model
}
\description{
Creates a NARX artificial neural network nonlinear
model (with autogressive and exogenous inputs)
}
