% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{annts}
\alias{annts}
\title{ANN-ts Model}
\usage{
annts(oy, nrn, afc)
}
\arguments{
\item{oy}{Vector with the lags in the autoregressive terms}

\item{nrn}{a vector (with dimension = number of hidden layers) with the number of neurons in each hidden layers}

\item{afc}{activation functions of the neurons (see ? layer_dense)}
}
\value{
Object representing the ANN-ts model
}
\description{
Creates a ANN-ts nonlinear model (with autogressive inputs)
}
