% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{arma}
\alias{arma}
\title{ARMA Model}
\usage{
arma(ny, ne)
}
\arguments{
\item{ny}{Number of autoregressive lags}

\item{ne}{Number of moving average lags}
}
\value{
Object representing an ARMA model
}
\description{
Creates an autoregressive moving average model
}
