% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{armax}
\alias{armax}
\title{ARMAX Model}
\usage{
armax(ny, nu, ne)
}
\arguments{
\item{ny}{Number of autoregressive lags}

\item{nu}{Number of input lags}

\item{ne}{Number of moving average lags}
}
\value{
Object representing an ARMAX model
}
\description{
Creates an autoregressive moving average with exogenous inputs model
}
