% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{arx}
\alias{arx}
\title{ARX Model}
\usage{
arx(ny, nu, p)
}
\arguments{
\item{ny}{Number of autoregressive lags}

\item{nu}{Number of input lags}
}
\value{
Object representing an ARX model
}
\description{
Creates an autoregressive with exogenous inputs model
}
