% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{baseRepr}
\alias{baseRepr}
\title{Base Representation}
\usage{
baseRepr(model)
}
\arguments{
\item{model}{A sysident model}
}
\value{
Boolean indicating if model is uninitialized
}
\description{
Generic function to print header
}
