% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcR2}
\alias{calcR2}
\title{Calculate R2}
\usage{
calcR2(real, est)
}
\arguments{
\item{real}{measured data}

\item{est}{predicted data}
}
\value{
R2
}
\description{
Calculate R2 (multiple correlation coefficient) for a predicted value
}
