% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{caret}
\alias{caret}
\title{caret-based Model}
\usage{
caret(oy, ou, method)
}
\arguments{
\item{oy}{Vector with the lags in the autoregressive terms}

\item{ou}{Vector with the lags in the exogenous terms}

\item{method}{model type (see https://topepo.github.io/caret/available-models.html)}
}
\value{
Object representing the caret-based model
}
\description{
Creates a NARX-caret based model
}
