% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crossco}
\alias{crossco}
\title{Calculate normalized cross-correlation of 2 signals
Constructed on the basis of the code provided by Norgaard, Ravn, Poulsen and Hansen: https://www.mathworks.com/matlabcentral/fileexchange/87-nnsysid}
\usage{
crossco(v, w, maxlag)
}
\description{
See Billings book, chapter 5
}
