% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estimate.ann}
\alias{estimate.ann}
\title{Estimate ANN-NARX model}
\usage{
\method{estimate}{ann}(model, Y, U, lr = 0.001, epochs = 100,
  batch_size = 32, verbose = 1)
}
\description{
Estimate ANN-NARX model
}
