% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estimate.annts}
\alias{estimate.annts}
\title{Estimate ANN-ts model}
\usage{
\method{estimate}{annts}(model, Y, lr = 0.001, epochs = 100,
  batch_size = 32, verbose = 1)
}
\description{
Estimate ANN-ts model
}
