% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estimate.caret}
\alias{estimate.caret}
\title{Estimate caret-NARX model}
\usage{
\method{estimate}{caret}(model, Y, U, trControl = NULL,
  tuneGrid = NULL, tuneLength = NULL)
}
\description{
Estimate caret-NARX model
}
