% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estimate}
\alias{estimate}
\title{Estimation function for NARMAX models}
\usage{
estimate(model, ...)
}
\arguments{
\item{model}{NARMAX model object}

\item{Y}{Measured data}

\item{U}{Corresponding input data}

\item{rho_p}{Threshold for ERR criterion: process terms}

\item{rho_n}{Threshold for ERR criterion: noise terms}
}
\value{
A model with selected terms and estimated parameters.
}
\description{
Proceed the estimation of the parameters using input/output data. ELS-FROLS algorithm is employed. Simpler versions apply for submodels.
}
