% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorizers.R
\name{factorize}
\alias{factorize}
\title{Factorize matrix}
\usage{
factorize(X, method = "mgs")
}
\arguments{
\item{X}{The matrix to be factorized}

\item{method}{Which method to use. Available methods are:
\describe{
  \item{"cgs"}{Classical Gram-Schimidt}
  \item{"mgs"}{Modified Gram-Schimidt}
}
It uses method = "mgs" by default.}
}
\value{
list(Q, A) Object containing factorization results
}
\description{
Decompose a matrix \eqn{X} into two simpler matrices
\eqn{Q} (Orthogonal columns) and \eqn{A} (Unit upper triangular), satisfying \eqn{X = Q * A}
}
