% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findTermIndexes}
\alias{findTermIndexes}
\title{Find term indexes}
\usage{
findTermIndexes(P, terms)
}
\arguments{
\item{P}{Regression matrix}

\item{terms}{Vector of terms (string format)}
}
\description{
Find term indexes in regression matrix
}
