% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frols.R
\name{frols}
\alias{frols}
\title{FROLS Algorithm}
\usage{
frols(P, Y, rho)
}
\arguments{
\item{P}{Matrix where each column is a potential feature}

\item{Y}{The target vector}

\item{rho}{Stop criteria (run while (1 - sum(EER)) > rho)}
}
\description{
Forward Regression with Orthogonal Least Squares
}
