% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generateRM.armaxModel}
\alias{generateRM.armaxModel}
\title{Generates a regression matrix}
\usage{
\method{generateRM}{armaxModel}(model, Y, U, E = NULL)
}
\arguments{
\item{model}{ARMAX Model}

\item{Y}{The target vector}

\item{U}{The input vector}

\item{E}{The error vector (can be NULL)}
}
\value{
Regression Matrix
}
\description{
Generates a regression matrix for an ARMAX Model
}
