% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generateRM.arxModel}
\alias{generateRM.arxModel}
\title{Generates a regression matrix}
\usage{
\method{generateRM}{arxModel}(model, Y, U, E = NULL)
}
\arguments{
\item{model}{ARX Model}

\item{Y}{The target vector}

\item{U}{The input vector}
}
\value{
Regression Matrix
}
\description{
Generates a regression matrix for an ARX Model
}
