% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{genRegMatrix.annts}
\alias{genRegMatrix.annts}
\title{Generates a regression matrix}
\usage{
\method{genRegMatrix}{annts}(model, Y, U = NULL, E = NULL)
}
\arguments{
\item{model}{ANN Model}

\item{Y}{The output data vector}
}
\value{
Regression matrix:
\describe{
 \item{Regression matrix with all terms}: y(t-1), ... contained in model information
}
}
\description{
Generates a regression matrix for a ANN-ts Model
}
