% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{genRegMatrix.arma}
\alias{genRegMatrix.arma}
\title{Generates a regression matrix}
\usage{
\method{genRegMatrix}{arma}(model, Y, U = NULL, E)
}
\arguments{
\item{model}{ARMAX Model}

\item{Y}{The target vector}

\item{U}{The input vector}

\item{E}{The error vector (can be NULL)}
}
\value{
Object containing:
\describe{
 \item{P}{Regression matrix with all terms}
 \item{Pp}{Regression matrix with only process terms}
 \item{Pnp}{Regression matrix without process terms}
}
}
\description{
Generates a regression matrix for an ARMAX Model
}
