% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{genRegMatrix.arx}
\alias{genRegMatrix.arx}
\title{Generates a regression matrix}
\usage{
\method{genRegMatrix}{arx}(model, Y, U, E = NULL)
}
\arguments{
\item{model}{ARX Model}

\item{Y}{The target vector}

\item{U}{The input vector}
}
\value{
Object containing:
\describe{
 \item{P}{Regression matrix with all terms}
 \item{Pp}{Regression matrix with only process terms}
 \item{Pnp}{Regression matrix without process terms}
}
}
\description{
Generates a regression matrix for an ARX Model
}
