% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{genRegMatrix.caret}
\alias{genRegMatrix.caret}
\title{Generates a regression matrix}
\usage{
\method{genRegMatrix}{caret}(model, Y, U, E = NULL)
}
\arguments{
\item{model}{any caret created model}

\item{Y}{The output data vector}

\item{U}{The input data vector}
}
\value{
Object containing:
\describe{
 \item{P}{Regression matrix with all terms}
 \item{Pp}{Regression matrix with only process terms}
 \item{Pnp}{Regression matrix without process terms}
}
}
\description{
Generates a regression matrix for a caret-based Model
}
