% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{genRegMatrix.narma}
\alias{genRegMatrix.narma}
\title{Generates a regression matrix}
\usage{
\method{genRegMatrix}{narma}(model, Y, U = NULL, E)
}
\arguments{
\item{model}{NARMA Model}

\item{Y}{The target vector}

\item{E}{The error vector}
}
\value{
Object containing:
\describe{
 \item{P}{Regression matrix with all terms}
 \item{Pp}{Regression matrix with only process terms}
 \item{Pnp}{Regression matrix without process terms}
}
}
\description{
Generates a regression matrix for a NARMA Model
}
