% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{genTarget.default}
\alias{genTarget.default}
\title{Generate target vector}
\usage{
\method{genTarget}{default}(model, Y)
}
\arguments{
\item{model}{Any model containing a $maxLag property}

\item{Y}{Original target vector}
}
\description{
Generate target vector based on maximum lag. Works with any model
}
