% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{M_spec}
\alias{M_spec}
\title{ploting signal spectrum}
\usage{
M_spec(u)
}
\arguments{
\item{u}{input vector to be analyzed}
}
\description{
performs U = fft(u)/sqrt(N), and plots 20*log10(abs(U)) for f in [0,0.5] (normalized)
}
\examples{
t=seq(from = 0, to = 400*2*pi, by = 0.01)
t = t[1:2^14]
u=sin(2*pi*t)
M_spec(u)
}
