% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model.armax}
\alias{model.armax}
\title{ARMAX Model}
\usage{
model.armax(na, nb, nc, p)
}
\arguments{
\item{na}{Number of autoregressive lags}

\item{nb}{Number of input lags}

\item{nc}{Number of moving average lags}

\item{p}{Autoregressive delay ??}
}
\value{
Object representing an ARMAX model
}
\description{
Creates an autoregressive moving average with exogenous inputs model
}
