% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model.arx}
\alias{model.arx}
\title{ARX Model}
\usage{
model.arx(na, nb, p)
}
\arguments{
\item{na}{Number of autoregressive lags}

\item{nb}{Number of input lags}

\item{p}{Autoregressive delay ??}
}
\value{
Object representing an ARX model
}
\description{
Creates an autoregressive with exogenous inputs model
}
