% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multisine}
\alias{multisine}
\title{Generate multisine excitation signal}
\usage{
multisine(N, cuoff)
}
\arguments{
\item{N}{number of samples}

\item{cutoff}{cutoff frequency (normalized)}
}
\value{
u, excitation signal
}
\description{
Ref.: Schoukens, Johan, Rik Pintelon, and Yves Rolain. Mastering system identification in 100 exercises. John Wiley & Sons, 2012.
}
