% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nar}
\alias{nar}
\title{NAR Model}
\usage{
nar(ny, nl)
}
\arguments{
\item{ny}{Number of autoregressive lags}

\item{nl}{Nonlinearity polinomial length}
}
\value{
Object representing a NAR model
}
\description{
Creates a nonlinear autoregressive model
}
