% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{narma}
\alias{narma}
\title{NARMA Model}
\usage{
narma(ny, ne, nl)
}
\arguments{
\item{ny}{Number of autoregressive lags}

\item{ne}{Number of moving average lags}

\item{nl}{Nonlinearity polinomial length}
}
\value{
Object representing a NARMA model
}
\description{
Creates a nonlinear autogressive moving average model
}
