% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{narmax}
\alias{narmax}
\title{NARMAX Model}
\usage{
narmax(ny, nu, ne, nl)
}
\arguments{
\item{ny}{Number of autoregressive lags}

\item{nu}{Number of input lags}

\item{ne}{Number of moving average lags}

\item{nl}{Nonlinearity polinomial length}
}
\value{
Object representing a NARMAX model
}
\description{
Creates a nonlinear autogressive moving average with exogenous inputs model
}
