% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{narx}
\alias{narx}
\title{NARX Model}
\usage{
narx(ny, nu, nl)
}
\arguments{
\item{ny}{Number of autoregressive lags}

\item{nu}{Number of input lags}

\item{nl}{Nonlinearity polinomial length}
}
\value{
Object representing a NARX model
}
\description{
Creates a nonlinear autoregressive with exogenous inputs model
}
