% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{print.armaxModel}
\alias{print.armaxModel}
\title{ARMAX model printer}
\usage{
\method{print}{armaxModel}(model)
}
\arguments{
\item{model}{ARMAX model}
}
\description{
Print basic info from ARMAX model
}
