% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{print.narmaxModel}
\alias{print.narmaxModel}
\title{NARMAX model printer}
\usage{
\method{print}{narmaxModel}(model)
}
\arguments{
\item{model}{NARMAX model}
}
\description{
Print basic info from NARMAX model
}
