% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{randnoise}
\alias{randnoise}
\title{Generate filtered random noise signal excitation signal}
\usage{
randnoise(N, cutoff)
}
\arguments{
\item{N}{number of samples}

\item{cutoff}{cutoff normalized frequency}
}
\value{
u, excitation signal
}
\description{
Filters N-normally distributed samples with a defined cut-off frequency

Ref.: Schoukens, Johan, Rik Pintelon, and Yves Rolain. Mastering system identification in 100 exercises. John Wiley & Sons, 2012.
}
