% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sayHello}
\alias{sayHello}
\title{Say Hello}
\usage{
sayHello(name, ...)
}
\arguments{
\item{name}{Name of the user to greet}
}
\description{
Prints hello to the console
}
\examples{
library(sysident)

sysident::sayHello('marcos')
}
