% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsetMatrix}
\alias{subsetMatrix}
\title{Subset matrix}
\usage{
subsetMatrix(mat, rows, cols)
}
\arguments{
\item{mat}{Target matrix}

\item{rows}{Rows subsetter}

\item{cols}{Cols subsetter}
}
\description{
Subset a matrix like mat[rows, cols]. Return a matrix even when the
subset generates a lower dimensional structure, i.e., never loose rownames or colnames.
}
