% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcR2}
\alias{calcR2}
\title{Calculate R2}
\usage{
calcR2(real, est)
}
\arguments{
\item{real}{measured data}

\item{est}{predicted data}
}
\value{
R2
}
\description{
Calculate R2 (multiple correlation coefficient) for a predicted value
}
\examples{
t=seq(from = 0, to = 400*2*pi, by = 0.01)
t = t[1:2^14]
u=sin(2*pi*t)
M_spec(u)
}
